// src/app/candidate/enums/sector.ts
export enum Sector { RESTAURANT = 'Hôtellerie-Restauration',
                     TRADE = 'Commerce',
                     MEDICALHEALTH = 'Santé-Médicale',
                     BUILDING = 'Bâtiment',
                     INDUSTRY = 'Industrie',
                     LOGISTICTRANSPORT = 'Transport-Logistique'
                    };